new: DUK-50: add sound pitch (#174)
* new: DUK-50: add frame rate controls to audio

* fix: DUK-50: check if object and component is valid before destroying

* chg: DUK-50: add frame rate test to audio on duk_sample

+++ duk_objects/include/duk_objects/objects.h
@@ -595,6 +595,9 @@ typename ObjectHandle<isConst>::ObjectsType* ObjectHandle<isConst>::objects() co
 
 template<bool isConst>
 void ObjectHandle<isConst>::destroy() const {
+    if (!valid()) {
+        return;
+    }
     m_objects->destroy_object(m_id);
 }
 
@@ -720,6 +723,9 @@ Id ComponentHandle<T, isConst>::id() const {
 
 template<typename T, bool isConst>
 ObjectHandle<isConst> ComponentHandle<T, isConst>::object() const {
+    if (!valid()) {
+        return {};
+    }
     return m_objects->object(m_ownerId);
 }
